+++ ChangeLog.md
@@ -17,6 +17,8 @@
 - Fixed a trashed raindrop, while still in trash, thinking its collection is
   the last collection it was in.
   ([#47](https://github.com/davep/braindrop/issues/47))
+- Made compact mode sticky.
+  ([#51](https://github.com/davep/braindrop/pull/51))
 
 ## v0.2.0